Added tank defence buff, defence actually does something now
Also, damage is now based on doubles rather than ints

+++ DungeonCrawler/src/engine/CircleDamageSource.java
@@ -4,11 +4,11 @@ import utility.Vector2D;
 
 public abstract class CircleDamageSource extends DamageSource {
 
-	public CircleDamageSource(Vector2D pos, int rad, int f, int d, boolean single, boolean p, int dam) {
+	public CircleDamageSource(Vector2D pos, int rad, int f, int d, boolean single, boolean p, double dam) {
 		super(new CircleHitbox(pos, rad), f, d, single, p, dam);
 	}
 	
-	public CircleDamageSource(Vector2D pos, int rad, int f, int d, boolean single, boolean p, int dam, StatusEffect e) {
+	public CircleDamageSource(Vector2D pos, int rad, int f, int d, boolean single, boolean p, double dam, StatusEffect e) {
 		super(new CircleHitbox(pos, rad), f, d, single, p, dam, e);
 	}
