+++ DungeonCrawler/src/engine/SwordDamageSource.java
@@ -9,7 +9,7 @@ public class SwordDamageSource extends DamageSource
 {
 
 	public SwordDamageSource(Vector2D position, int radius, int startAngle,
-			int angle, int d, boolean p, int dam)
+			int angle, int d, boolean p, double dam)
 	{
 		super(new SwordHitbox(position, radius, startAngle, angle), 0, d,
 				false, p, dam);
